+++ Dyalect/Compiler/CodeWriter.cs
@@ -226,6 +226,30 @@ namespace Dyalect.Compiler
             Emit(new Op(OpCode.Tag, idx));
         }
 
+        public void SetMember(TypeHandle type)
+        {
+            if (type.IsStandard)
+                Emit(new Op(OpCode.SetMemberT, type.TypeId));
+            else
+                Emit(new Op(OpCode.SetMember, type.TypeId));
+        }
+
+        public void SetMemberS(TypeHandle type)
+        {
+            if (type.IsStandard)
+                Emit(new Op(OpCode.SetMemberST, type.TypeId));
+            else
+                Emit(new Op(OpCode.SetMemberS, type.TypeId));
+        }
+
+        public void TypeCheck(TypeHandle type)
+        {
+            if (type.IsStandard)
+                Emit(new Op(OpCode.TypeCheckT, type.TypeId));
+            else
+                Emit(new Op(OpCode.TypeCheck, type.TypeId));
+        }
+
         public void FunPrep(int argCount) => Emit(new Op(OpCode.FunPrep, argCount));
         public void FunArgIx(int index) => Emit(new Op(OpCode.FunArgIx, index));
         public void FunArgNm(string name) => Emit(new Op(OpCode.FunArgNm, IndexString(name)));
@@ -240,9 +264,6 @@ namespace Dyalect.Compiler
         public void Brfalse(Label lab) => Emit(OpCode.Brfalse, lab);
         public void Brterm(Label lab) => Emit(OpCode.Brterm, lab);
         public void Briter(Label lab) => Emit(OpCode.Briter, lab);
-        public void TypeCheck(int type) => Emit(new Op(OpCode.TypeCheck, type));
-        public void SetMember(int type) => Emit(new Op(OpCode.SetMember, type));
-        public void SetMemberS(int type) => Emit(new Op(OpCode.SetMemberS, type));
         public void GetMember(int nameId) => Emit(new Op(OpCode.GetMember, nameId));
         public void HasMember(int nameId) => Emit(new Op(OpCode.HasMember, nameId));
         public void RunMod(int code) => Emit(new Op(OpCode.RunMod, code));
@@ -252,6 +273,7 @@ namespace Dyalect.Compiler
         public void HasField(string field) => Emit(new Op(OpCode.HasField, IndexString(field)));
         public void Start(Label lab) => Emit(OpCode.Start, lab);
         public void Fail(DyErrorCode code) => Emit(new Op(OpCode.FailSys, (int)code));
+        public void NewType(int typeId) => Emit(new Op(OpCode.NewType, typeId));
 
         public void End() => Emit(Op.End);
         public void Yield() => Emit(Op.Yield);