+++ Dyalect/Parser/Model/DTypeDeclaration.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using System.Text;
+﻿using System.Text;
 
 namespace Dyalect.Parser.Model
 {
@@ -12,18 +11,10 @@ namespace Dyalect.Parser.Model
 
         public string Name { get; set; }
 
-        public List<DFieldDeclaration> Fields { get; } = new List<DFieldDeclaration>();
-
-        public DNode Body { get; set; }
-
         internal override void ToString(StringBuilder sb)
         {
             sb.Append("type ");
             sb.Append(Name);
-            sb.Append(' ');
-            sb.Append('{');
-            Fields.ToString(sb);
-            sb.Append('}');
         }
     }
 }